feat: Refactored utils for Sparkle component; added back missing exports in index.ts
