+++ web/libs/ui/src/assets/icons/sparkle.module.scss
@@ -33,9 +33,4 @@
   100% {
     transform: rotate(90deg);
   }
-}
-
-.sparkles-area-overlay {
-  /* Controls the color of the area overlay for showArea */
-  color: var(--color-accent-primary-base, #0057FF);
 }
