+++ web/libs/ui/src/assets/icons/sparkle.tsx
@@ -1,0 +1,21 @@
+import React from "react";
+import styles from "./sparkle.module.scss";
+
+export interface SparkleProps {
+  size: number;
+  color: string;
+  style: React.CSSProperties;
+  className?: string;
+}
+
+export const Sparkle: React.FC<SparkleProps> = ({ size, color, style, className }) => (
+  <span style={style} className={[styles.sparkle, className].filter(Boolean).join(" ")}>
+    <svg width={size} height={(size * 5) / 6} viewBox="0 0 6 5" fill="none">
+      <title>Sparkle icon</title>
+      <path
+        d="M2.96504 4.88359C2.92376 4.82938 2.88744 4.64775 2.81478 4.2845C2.76575 4.03932 2.74123 3.91672 2.69733 3.8104C2.57741 3.51993 2.3467 3.28922 2.05624 3.1693C1.94991 3.1254 1.82732 3.10089 1.58214 3.05185C1.21888 2.9792 1.03725 2.94287 0.983048 2.90159C0.827855 2.7834 0.827855 2.54989 0.983048 2.4317C1.03725 2.39042 1.21888 2.3541 1.58213 2.28145C1.82731 2.23241 1.94991 2.20789 2.05624 2.16399C2.3467 2.04407 2.57741 1.81336 2.69733 1.5229C2.74123 1.41657 2.76575 1.29398 2.81478 1.04881C2.88743 0.685548 2.92376 0.503909 2.96504 0.449707C3.08323 0.294515 3.31674 0.294515 3.43493 0.449707C3.47621 0.503909 3.51254 0.685535 3.58519 1.04879C3.63422 1.29397 3.65874 1.41657 3.70264 1.5229C3.82256 1.81336 4.05327 2.04407 4.34374 2.16399C4.45006 2.20789 4.57266 2.23241 4.81784 2.28144C5.18109 2.35409 5.36272 2.39042 5.41693 2.4317C5.57212 2.54989 5.57212 2.7834 5.41693 2.90159C5.36272 2.94287 5.1811 2.9792 4.81784 3.05185C4.57266 3.10089 4.45006 3.1254 4.34374 3.1693C4.05327 3.28922 3.82256 3.51993 3.70264 3.8104C3.65874 3.91672 3.63423 4.03931 3.58519 4.2845C3.51254 4.64775 3.47621 4.82938 3.43493 4.88359C3.31674 5.03878 3.08323 5.03878 2.96504 4.88359Z"
+        fill={color}
+      />
+    </svg>
+  </span>
+);
