+++ web/libs/ui/src/index.ts
@@ -19,6 +19,8 @@ export * from "./assets/icons";
 export * from "./lib/simple-card";
 export * from "./lib/space/space";
 export * from "./lib/spinner/spinner";
-export * from "./lib/sparkles";
+export * from "./lib/sparkles/sparkles";
+export * from "./lib/popover/popover";
+export * from "./lib/auto-sizer-table/auto-sizer-table";
 
 export * from "./shadcn";
